Moves an anonymous calls into a packaged one.
